server/sm-pipeline-cdk/functions/requirements.txt
aws-lambda-powertools[Logger]
python-jose[cryptography]
chevron
pandas
numpy